Config improvements
