--- .env.aad.cherrytwist.api.default
@@ -1,4 +1,0 @@
-AUTH_AAD_CHERRYTWIST_API_APP_ID=505041fc-fca2-4a74-88ee-6d50a6417e38
-AUTH_AAD_MSGRAPH_API_SECRET=5S2p55u-OuSaWu~6igKgmJZ2q..v5cL_P5
-AUTH_AAD_MSGRAPH_API_SCOPE='Directory.AccessAsUser.All Directory.ReadWrite.All User.ReadWrite.All'
-AUTH_AAD_UPN_DOMAIN=playgroundcherrytwist.onmicrosoft.com
