--- .env.aad.cherrytwist.client.default
@@ -1,3 +1,0 @@
-AUTH_AAD_CLIENT_APP_ID=feb6d8b1-8cc2-4e7d-a419-ad7b544b0832
-AUTH_AAD_CHERRYTWIST_API_SCOPE=api://505041fc-fca2-4a74-88ee-6d50a6417e38/.default
-AUTH_AAD_CLIENT_LOGIN_REDIRECT_URI=http://localhost:3000
