--- .env.default
@@ -1,14 +1,0 @@
-WAIT_HOSTS=localhost:3306
-WAIT_HOSTS_TIMEOUT=300
-WAIT_SLEEP_INTERVAL=30
-WAIT_HOST_CONNECT_TIMEOUT=30
-GRAPHQL_ENDPOINT_PORT=4000
-AUTH_AAD_TENANT=524e761c-d162-4fdf-ab43-2855246d986c
-AUTH_ENABLED=true
-
-AUTH_BOOTSTRAP_PATH=
-TEMPLATES_BOOTSTRAP_PATH
-
-
-
-COMPOSE_PROJECT_NAME=cherrytwist-server
