+++ cherrytwist.yml
@@ -31,8 +31,28 @@ service:
   authorisationBootstrapPath: ${AUTH_BOOTSTRAP_PATH}
 aad_ropc:
   clientID: ${AUTH_AAD_CLIENT_APP_ID}
-  clientSecret: ${AUTH_AAD_CLIENT_APP_SECRET}
   scope: ${AUTH_AAD_CHERRYTWIST_API_SCOPE}
   tenant: ${AUTH_AAD_TENANT}
   username: ${AUTH_AAD_USER_UPN}
   password: ${AUTH_AAD_USER_PASSWORD}
+aad:
+  identityMetadata: ${AUTH_AAD_IDENTITY_METADATA}:https://login.microsoftonline.com/524e761c-d162-4fdf-ab43-2855246d986c/v2.0/.well-known/openid-configuration
+  clientID: ${AUTH_AAD_CHERRYTWIST_API_APP_ID}:505041fc-fca2-4a74-88ee-6d50a6417e38
+  validateIssuer: true
+  passReqToCallback: true
+  issuer: ${AUTH_AAD_ISSUER}:https://login.microsoftonline.com/524e761c-d162-4fdf-ab43-2855246d986c/v2.0
+  audience: ${AUTH_AAD_CHERRYTWIST_API_APP_ID}:505041fc-fca2-4a74-88ee-6d50a6417e38
+  allowMultiAudiencesInToken: false
+  loggingLevel: ${AUTH_AAD_LOGGING_LEVEL}:Error
+  scope: ['Cherrytwist-GraphQL']
+  loggingNoPII: false
+  upnDomain: ${AUTH_AAD_UPN_DOMAIN}:playgroundcherrytwist.onmicrosoft.com
+  tenant: ${AUTH_AAD_TENANT}
+  ropc:
+    clientSecret: ${AUTH_AAD_CLIENT_APP_SECRET}
+    username: ${AUTH_AAD_USER_UPN}
+    password: ${AUTH_AAD_USER_PASSWORD}
+  client:
+    clientId: ${AUTH_AAD_CLIENT_APP_ID}:feb6d8b1-8cc2-4e7d-a419-ad7b544b0832
+    authority: ${AUTH_AAD_AUTHORITY}:https://login.microsoftonline.com/524e761c-d162-4fdf-ab43-2855246d986c
+    redirectUri: ${AUTH_AAD_CLIENT_LOGIN_REDIRECT_URI}:http://localhost:3000
