+++ src/common/enums/aad.logging.level.ts
@@ -1,0 +1,5 @@
+export enum AAD_LOGGING_LEVEL {
+  Error = 'error',
+  Warning = 'warn',
+  Info = 'info',
+}
