+++ src/config/index.ts
@@ -1,4 +1,2 @@
-export * from './aad.client.config';
-export * from './aad.config';
 export * from './typeorm.cli.config';
 export * from './winston.config';
